Updated some dependencies
Some of the dependencies needed to be updated as they were old, and we
needed to be able to move ahead with other libraries, and they had moved
on. So this was primarily about getting these components up to a current
level, so that this could be included in other projects.